+++ UI/index.html
@@ -132,7 +132,7 @@
     <article>
         <div id="note-div">
     
-            <a href="" class="buttonnote" id="add_new">Add Note</a>
+            <a href="javascript:;" class="buttonnote" id="add_new">Add Note</a>
             <div id="board">
               
             </div>
