+++ cgi_bin/inputselector.py
@@ -2,118 +2,120 @@ import cgi, cgitb, json, pymysql.cursors
 
 
 
-def setup_db_connect():
-	dbconn = pymysql.connect(host='localhost',
-				 user='root',
-				 password='defaultmysqlpassword',
-				 db='db',
-				 charset='utf8mb4',
-				 cursorclass=pymysql.cursors.DictCursor)
+#def setup_db_connect():
+#	global dbconn
+#	dbconn = pymysql.connect(host='localhost',
+#				 user='root',
+#				 password='',
+#				 db='db',
+#				 charset='utf8mb4',
+#				 cursorclass=pymysql.cursors.DictCursor)
 
 
-def add_entry(field_storage):
-	try:	
-		format = field_storage['format'].getvalue #note or calendar?
-		text = field_storage['text'].getvalue
-		user = field_storage['user'].getvalue
-		if (format == note):
-			try:
-				with connection.cursor() as cursor:
-					cmd = "INSERT INTO %s (username, text) VALUES (%s, %s)" % (format,user,text) #s's represent args
-					cursor.execute(cmd); #Run it
-					dbconn.commit(); #change needs to be committed
-			except:
-				raise Exception()
+#def add_entry(field_storage):
+#	try:	
+#		format = field_storage['format'].getvalue #note or calendar?
+#		text = field_storage['text'].getvalue
+#		user = field_storage['user'].getvalue
+#		if (format == note):
+#			try:
+#				with connection.cursor() as cursor:
+#					cmd = "INSERT INTO %s (username, text) VALUES (%s, %s)" % (format,user,text) #s's represent args
+#					cursor.execute(cmd); #Run it
+#					dbconn.commit(); #change needs to be committed
+#			except:
+#				raise Exception()
 		
-		if (format == calendar):
-			date = field_storage['date'].getvalue #TODO Going to have to format this so SQL is ok with it
-			try:
-				with connection.cursor() as cursor:
-					cmd = "INSERT INTO %s (username, text, date) VALUES (%s, %s, %s)"% (format, user, text, date) #s's represent args
-					cursor.execute(cmd); #Run it
-					dbconn.commit() #change needs to be committed
-			except:
-				raise Exception()
+#		if (format == calendar):
+#			date = field_storage['date'].getvalue #TODO Going to have to format this so SQL is ok with it
+#			try:
+#				with connection.cursor() as cursor:
+#					cmd = "INSERT INTO %s (username, text, date) VALUES (%s, %s, %s)"% (format, user, text, date) #s's represent args
+#					cursor.execute(cmd); #Run it
+#					dbconn.commit() #change needs to be committed
+#			except:
+#				raise Exception()
 
-	except:
-		return_error('add entry failed')
-		return
-	return_success('add entry success') #Need to grab the id of the new entry and send it back to the frontend
+#	except:
+#		return_error('add entry failed')
+#		return
+#	return_success('add entry success') #Need to grab the id of the new entry and send it back to the frontend
 
 
-def edit_entry(field_storage):
+#def edit_entry(field_storage):
 
-	try:
-		format = field_storage['format'].getvalue #note or calendar?
-		id = field_storage()['id'].getvalue	#value is what will be used to figure out what to update
+#	try:
+#		format = field_storage['format'].getvalue #note or calendar?
+#		id = field_storage()['id'].getvalue	#value is what will be used to figure out what to update
 	
-		try:
-			with connection.cursor() as cursor:
-				text = field_storage['text'].getvalue
-				cmd = "UPDATE %s SET text = %s WHERE id = %s" % (format, text, id)
-				cursor.execute(cmd)
-				dbconn.commit()		
-		except KeyError:
-			raise Exception()
+#		try:
+#			with connection.cursor() as cursor:
+#				text = field_storage['text'].getvalue
+#				cmd = "UPDATE %s SET text = %s WHERE id = %s" % (format, text, id)
+#				cursor.execute(cmd)
+#				dbconn.commit()		
+#		except KeyError:
+#			raise Exception()
 		
-		try:
-			with connection.cursor() as cursor:
-				date = field_storage['date'].getvalue
-				cmd = 'UPDATE %s SET date = %s WHERE id = %s' % (format, date, id)
-				cursor.execute(cmd)
-				dbconn.commit()
-		except KeyError:
-			raise Exception()
+#		try:
+#			with connection.cursor() as cursor:
+#				date = field_storage['date'].getvalue
+#				cmd = 'UPDATE %s SET date = %s WHERE id = %s' % (format, date, id)
+#				cursor.execute(cmd)
+#				dbconn.commit()
+#		except KeyError:
+#			raise Exception()
 
-	except:
-		return_error('edit failed')
-		return
+#	except:
+#		return_error('edit failed')
+#		return
 	
-	return_success('edit success')
+#	return_success('edit success')
 
 
-def delete_entry(field_storage):
-	try:
-		format = field_storage['format'].getvalue
-		id = field_storage['id'].getvalue	
-		with connection.cursor() as cursor:
-			cmd = "DELETE FROM %s WHERE id= %s" % (format, id) #s's represent args
-			cursor.execute(cmd); #Run it
-			dbconn.commit(); #change needs to be committed
-	except:
-		return_error('delete failed')
+#def delete_entry(field_storage):
+#	try:
+#		format = field_storage['format'].getvalue
+#		id = field_storage['id'].getvalue	
+#		with connection.cursor() as cursor:
+#			cmd = "DELETE FROM %s WHERE id= %s" % (format, id) #s's represent args
+#			cursor.execute(cmd); #Run it
+#			dbconn.commit(); #change needs to be committed
+#	except:
+#		return_error('delete failed')
 
 
-def return_error(message):
-		print(json.JSONEncoder().encode({'error',message}))
+#def return_error(message):
+#		print(json.JSONEncoder().encode({'error',message}))
 
 
-def return_success(message):
-		print(json.JSONEncoder().encode({'success',message}))
+#def return_success(message):
+#		print(json.JSONEncoder().encode({'success',message}))
 
 
-def main():
-	cgitb.enable()
-
-	print("Content-Type: application/json")
-	print()
-
-	setup_db_connect()
+cgitb.enable()
 	
-	field_storage = cgi.FieldStorage()
+cgi.test()
 
-	_type = field_storage['type'].getvalue  #This queries the field storage for the raw string associated with the key
+#	print("Content-Type: application/json")
+#	print()
 
-	if _type == 'add':
-		add_entry(field_storage)
+	#setup_db_connect()
 	
-	if _type == 'edit':
-		edit_entry(field_storage)
-
-	if _type == 'delete':
-		delete_entry(field_storage)
+	#field_storage = cgi.FieldStorage()
 	
-	dbconn.close()
+	#print(json.JSONEncoder().encode(field_storage))
 
-if __name__ == '__main__':
-	main()
+	#_type = field_storage['type'].getvalue  #This queries the field storage for the raw string associated with the key
+
+	#if _type == 'add':
+	#	add_entry(field_storage)
+	
+	#if _type == 'edit':
+	#	edit_entry(field_storage)
+
+	#if _type == 'delete':
+	#	delete_entry(field_storage)
+	
+	#dbconn.close()
+
