+++ cgi_bin/setupdb.sql
