+++ cgi_bin/test.py
@@ -4,11 +4,10 @@ import cgi, cgitb, json
 
 cgitb.enable()
 
-print("Content-Type: application/json")
-print()
+cgi.test()
 
-arg_input = cgi.FieldStorage()['input'].value
-response = 'Bounced off server: ' + str(arg_input)
-print(json.JSONEncoder().encode(response))
+#print("Content-Type: application/json")
+#print()
 
+#print(json.JSONEncoder().encode("script ran"))
